Modified plot_data function
Modified plot_data function to take two inputs; the "grouped_data" as before, and additionally, "data_type". The "data_type" is a string that is either 'income' or 'expenses'; specifying this let's us dynamically name the bar plot to match the type of data being plotted (income title for income plot, expenses title for expenses plot).