Updated net income function
Updated the plot_net_income function and fixed a bug so that now it displays my net income for ALL months, not just the months during which I made an income. This means that if I earned $0 in a month, but spent $1'000, it will show my net income as –$1'000 (whereas previously it would have just not drawn a bar for that month, because I didn't make any income that month).