Finance Report PDF
Created a new Python file called "generate_report.py" to experiment and learn how to generate PDF reports with Python. The current goal is to take all the plots generated by the "analyze_finances.py", and create a PDF report containing those plots. The ultimate goal is to automate the entire process of inputting income/expenses, and let the code run by itself to classify the expenses, analyze finances, generate a report, and email it to me on a regular basis.